Modernize wrapper for _BaseColumnViewBase.
