+++ python/lsst/afw/table/__init__.py
@@ -22,7 +22,7 @@
 from ._table import *
 from ._aliasMap import *
 from ._schema import *
-from .baseColumnView import *
+from ._baseColumnView import *
 from .base import *
 from .idFactory import *
 from .aggregates import *
